Implemented async register dialog
