+++ client/async/UserRegisterBox.js
@@ -1,0 +1,1 @@
+export { default } from '../components/UserRegisterBox'
