+++ client/components/Dialog/Dialog.js
@@ -6,12 +6,23 @@ import { compose } from 'recompose'
 import Overlay from './Overlay'
 import UniversalComponent from '../UniversalComponent'
 
-import * as actions from './actions'
+import {
+  DIALOG_IDS,
+  openDialog as openDialogAction,
+  closeDialog as closeDialogAction,
+} from './actions'
+
+const dialogComponentNames = {
+  [DIALOG_IDS.LOGIN]: 'UserLoginBox',
+  [DIALOG_IDS.REGISTER]: 'UserRegisterBox',
+}
 
 const getDialogForId = (dialogId) => {
-  if (!dialogId) return null
-
-  const name = 'UserLoginBox'
+  const name = dialogComponentNames[dialogId]
+  if (!name) {
+    console.error(`No dialog component for dialog ${dialogId} found`)
+    return null
+  }
   return <UniversalComponent name={name} />
 }
 
@@ -34,8 +45,8 @@ const mapStateToProps = state => ({
 })
 
 const mapDispatchToProps = dispatch => ({
-  openDialog: id => dispatch(actions.openDialog(id)),
-  closeDialog: () => dispatch(actions.closeDialog()),
+  openDialog: id => dispatch(openDialogAction(id)),
+  closeDialog: () => dispatch(closeDialogAction()),
 })
 
 const enhance = compose(
