--- client/pages/UserRegister/FormWrapper.js
@@ -1,8 +1,0 @@
-import styled from 'styled-components'
-
-const FormWrapper = styled.div`
-  width: 320px;
-  margin-left: auto;
-  margin-right: auto;
-`
-export default FormWrapper
