+++ client/pages/UserRegister/Wrapper.js
@@ -2,7 +2,8 @@ import styled from 'styled-components'
 
 const Wrapper = styled.div`
   margin-top: 50px;
-  text-align: center;
+  display: flex;
+  justify-content: center;
 `
 
 export default Wrapper
